move getIconByTheme
